chore: use GUID to register component information of DynamicBone
