--- Editor/ExternalLibraryAccessor.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 644e19c4d8e44e3b82e368773cbb72f9
-timeCreated: 1693815396
